+++ src/main/java/org/stellar/sdk/responses/PathResponse.java
@@ -10,8 +10,10 @@ import java.util.ArrayList;
 /**
  * Represents path response.
  * @see <a href="https://www.stellar.org/developers/horizon/reference/resources/path.html" target="_blank">Path documentation</a>
- * @see org.stellar.sdk.requests.PathsRequestBuilder
- * @see org.stellar.sdk.Server#paths()
+ * @see org.stellar.sdk.requests.StrictReceivePathsRequestBuilder
+ * @see org.stellar.sdk.requests.StrictSendPathsRequestBuilder
+ * @see org.stellar.sdk.Server#strictReceivePaths()
+ * @see org.stellar.sdk.Server#strictSendPaths()
  */
 public class PathResponse extends Response {
   @SerializedName("destination_amount")